Fixes breaking when Identity Provider does not provide a new id_token
According to section 12.2 of the OpenID Connect Core 1.0 specifications:
"Upon successful validation of the Refresh Token, the response body is the
Token Response of Section 3.1.3.3 except that it might not contain an id_token."

This means we cannot make the assumption that we get a new id token on refresh.
If we don't get a new ID Token, let's just reuse the old one, updating the exp
with the expires_in of the new token.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>