Make error strings returned by validate_token have bool() as False
I did not apply the "principle of least surprise" with the validate_token API,
and users had to do explicit "if valid is True:", because the error strings
returned were also evaluating to True.
This change returns ErrStr objects, which evaluate to False, but otherwise are
the exact same as normal string objects.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>